% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_flow.r
\name{dev_flow}
\alias{dev_flow}
\title{Send the graphics contents of the current device to a Flowdock flow}
\usage{
dev_flow(..., file = "Rplot", flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW",
  NULL), tags = "", flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{...}{other arguments passed into png device}

\item{file}{prefix for filenames (defaults to \code{plot})}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Send the graphics contents of the current device to a Flowdock flow
}

