% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-file.r
\name{flow_file}
\alias{flow_file}
\title{Upload a file to a flow}
\usage{
flow_file(fil = NULL, flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW", NULL),
  tags = "", flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{fil}{path to file to upload}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Upload a file to a flow
}

