% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-msg.r
\name{flow_msg}
\alias{flow_msg}
\title{Post a message to a flow}
\usage{
flow_msg(message = NULL, flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW", NULL),
  tags = "", flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{message}{message to send}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Post a message to a flow
}

