% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-r.r
\name{flow_r}
\alias{flow_r}
\title{Post R objects to Flowdock}
\usage{
flow_r(..., flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW", NULL), tags = "",
  flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{...}{one or more R objects}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Post R objects to Flowdock
}

