% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowdockr-package.R
\docType{package}
\name{flowdockr}
\alias{flowdockr}
\alias{flowdockr-package}
\title{Tools to Work with Flowdock}
\description{
Heavily based on \href{http://github.com/hrbrmstr/slackr}{slackr}.
}
\author{
Bob Rudis (@hrbrmstr)
}

