% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-flow-info.r
\name{get_flow_info}
\alias{get_flow_info}
\title{Retrieve flow metadata}
\usage{
get_flow_info(flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW", NULL),
  flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Retrieve flow metadata
}

