% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflow.r
\name{ggflow}
\alias{ggflow}
\title{Send a rendered ggplot object to Flowdock}
\usage{
ggflow(plot = last_plot(), file = "ggplot",
  flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW", NULL), tags = "", scale = 1,
  width = par("din")[1], height = par("din")[2], units = c("in", "cm",
  "mm"), dpi = 300, limitsize = TRUE,
  flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"), ...)
}
\arguments{
\item{plot}{ggplot object}

\item{file}{prefix for filenames (defaults to \code{ggplot})}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{scale}{scaling factor}

\item{width}{width (defaults to the width of current plotting window)}

\item{height}{height (defaults to the height of current plotting window)}

\item{units}{units for width and height when either one is explicitly specified (in, cm, or mm)}

\item{dpi}{dpi to use for raster graphics}

\item{limitsize}{when TRUE (the default), ggsave will not save images larger than 50x50 inches,
to prevent the common error of specifying dimensions in pixels.}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}

\item{...}{other arguments passed to graphics device}
}
\value{
parsed call result (invisibly)
}
\description{
Send a rendered ggplot object to Flowdock
}

