% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-flow.r
\name{save_flow}
\alias{save_flow}
\title{Save R objects to an RData file on Flowdock}
\usage{
save_flow(..., file = "flowr", flow = Sys.getenv("FLOWDOCK_DEFAULT_FLOW",
  NULL), tags = "", flowdock_api_key = Sys.getenv("FLOWDOCK_PAT"))
}
\arguments{
\item{...}{objects to store in the R data file}

\item{file}{filename (without extension) to use}

\item{flow}{parameterized flow name (i.e. "\code{this-flow}")}

\item{tags}{vector of tags for the flow}

\item{flowdock_api_key}{used to authorize you to Flowdoc. You should store this
in \code{.Renviron} or some other moderately secure place. You can override
the pick from the environment by passing it in here.}
}
\value{
parsed call result (invisibly)
}
\description{
Save R objects to an RData file on Flowdock
}

